- some restructuring - add flags for run - further completion of readme
